Specify content suitable for public distribution.
The whitelist of suitable content is intentionally not exhaustive;
instead, an clear exception mechanism is described.  This keeps the
policy short and hopefully leads to quick adjudication of exceptions.

Additionally, the phrase "and there is any question about its
suitability" is included to grandfather in content such as "deps"
packages without needing to spell out every last exception.